refactor(PurchaseAmount): amount를 long에서 int로 변경

+++ src/main/java/lotto/domain/PurchaseAmount.java
@@ -7,31 +7,31 @@ public class PurchaseAmount {
     private static final int MINIMUM = 1_000;
     private static final int MAXIMUM = 100_000_000;
     private static final int LOTTO_PRICE_UNIT = 1_000;
-    private final long amount;
+    private final int amount;
 
-    public PurchaseAmount(long amount) {
+    public PurchaseAmount(int amount) {
         validate(amount);
         this.amount = amount;
     }
 
-    private void validate(long amount) {
+    private void validate(int amount) {
         validateRange(amount);
         validateUnit(amount);
     }
 
-    private void validateRange(long amount) {
+    private void validateRange(int amount) {
         if (amount < MINIMUM || amount > MAXIMUM) {
             throw new IllegalArgumentException(PurchaseAmountExceptionMessage.OUT_OF_AMOUNT_RANGE.getError());
         }
     }
 
-    private void validateUnit(long amount) {
+    private void validateUnit(int amount) {
         if (amount % LOTTO_PRICE_UNIT != 0) {
             throw new IllegalArgumentException(PurchaseAmountExceptionMessage.INVALID_AMOUNT_UNIT.getError());
         }
     }
 
-    public long getPurchaseLottoCount() {
+    public int getPurchaseLottoCount() {
         return amount / LOTTO_PRICE_UNIT;
     }
 }
